odata: narrow EDM int mapping — Int32 and below to IntegerType
Revert the LongType-everywhere change. The framework's parse_value
supports both IntegerType and LongType, so there's no need to widen
Edm.Byte/SByte/Int16/Int32 all the way to LongType; they fit in
Int32 by definition and IntegerType is the more accurate Spark type.
Only Edm.Int64 still maps to LongType.

ShortType / ByteType remain unused (the framework doesn't support
them).

+++ tests/unit/sources/odata/test_odata_lakeflow_connect.py
@@ -19,7 +19,7 @@ import responses
 
 from databricks.labs.community_connector.sources.odata import ODataLakeflowConnect
 from databricks.labs.community_connector.sources.odata.odata import _odata_literal
-from pyspark.sql.types import LongType, StringType, TimestampType
+from pyspark.sql.types import IntegerType, StringType, TimestampType
 from tests.unit.sources.test_suite import LakeflowConnectTests
 
 
@@ -126,7 +126,7 @@ def test_get_table_schema_maps_edm_types():
     names = [f.name for f in schema.fields]
     types = [type(f.dataType).__name__ for f in schema.fields]
     assert names == ["Id", "Name", "ModifiedAt"]
-    assert types == ["LongType", "StringType", "TimestampType"]
+    assert types == ["IntegerType", "StringType", "TimestampType"]
     assert schema.fields[0].nullable is False
